mandelbrot -- fixes

+++ prs_mandelbrot/src/lib.rs
@@ -131,7 +131,7 @@ fn write_image(filename: &str, pixels: &[u8], bounds: (usize, usize))
   // is equal to:
   let output = File::create(filename)?;
   let encoder = PNGEncoder::new(output);
-  encoder.encode(&pixels, bounds.0, bounds.1, ColorType::Gray(8))?;
+  encoder.encode(&pixels, bounds.0 as u32, bounds.1 as u32, ColorType::Gray(8))?;
 
   Ok(())
 }
